moved TalleMenu under browsingtable
