+++ src/dbtarzan/gui/MainGUI.scala
@@ -15,6 +15,7 @@ import akka.actor.ActorRef
 import dbtarzan.gui.util.JFXUtil
 import dbtarzan.gui.config.connections.ConnectionEditorStarter
 import dbtarzan.config.password.{EncryptionKey, VerificationKey, PasswordEncryption}
+import dbtarzan.gui.browsingtable.TableMenu
 import dbtarzan.gui.config.global.GlobalEditorStarter
 import dbtarzan.types.ConfigPath
 import dbtarzan.messages.Logger
